pyproject.toml
[project]
name = "genius-mcp"
version = "0.1.0"
description = "MCP server for Genius lyrics and annotations"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Genius MCP", email = "" }
]
license = { text = "MIT" }

dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=0.21.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
]

[project.scripts]
genius-mcp = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []

